- refector lifecycle status
