+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifeCycle/requestStatus/RequestStatusAvailability.java
@@ -1,31 +1,38 @@
-package de.samply.bbmri.negotiator.util.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
 
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
 import org.jooq.tools.json.JSONObject;
 import org.jooq.tools.json.JSONParser;
 import org.jooq.tools.json.ParseException;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusAccessConditions implements RequestStatus {
+public class RequestStatusAvailability implements RequestStatus {
 
     private String status = null;
-    private String statusType = "accessConditions";
-    private String statusText = "Access Condition indicated for collection.";
+    private final String statusType = "availability";
+    private String statusText = "Collection has availability not specified yet.";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("not_interrested", "selectAndAcceppt");
+    private final List<String> allowedNextStatus = Arrays.asList("not_interrested", "indicateAccessConditions");
+    private final List<String> allowedNextStatusBiobanker = new ArrayList<String>();
+    private final List<String> allowedNextStatusResearcher = Arrays.asList("notselected.watingForResponse", "abandoned.not_interrested");
 
-    private List allowedNextStatusBiobanker = Arrays.asList("notselected.watingForResponseFromResearcher", "abandoned.not_interrested");
-
-    private List allowedNextStatusResearcher = Arrays.asList("accepptConditions.selectAndAcceppt", "abandoned.not_interrested");
-
-    public RequestStatusAccessConditions(CollectionRequestStatusDTO collectionRequestStatusDTO) {
+    public RequestStatusAvailability(CollectionRequestStatusDTO collectionRequestStatusDTO) {
         statusDate = collectionRequestStatusDTO.getStatusDate();
         status = collectionRequestStatusDTO.getStatus();
-        if(status.equals("indicateAccessConditions")) {
-            statusText = "Access Condition: " + getStatusTextFromJson(collectionRequestStatusDTO.getStatusJson(), "indicateAccessConditions");
+        allowedNextStatusBiobanker.add("notselected.notselected");
+        if(status.equalsIgnoreCase("sample_data_available_accessible")) {
+            allowedNextStatusBiobanker.add("accessConditions.indicateAccessConditions");
+        }
+        allowedNextStatusBiobanker.add("abandoned.not_interrested");
+        if(status.equals("sample_data_available_accessible")) {
+            String numberAvaiableSamples = getStatusTextFromJson(collectionRequestStatusDTO.getStatusJson(), "numberAvaiableSamples");
+            if(numberAvaiableSamples != null && numberAvaiableSamples.length() > 0) {
+                statusText = "Number of avaiable Samples: " + numberAvaiableSamples;
+            }
         }
     }
 
@@ -46,7 +53,10 @@ public class RequestStatusAccessConditions implements RequestStatus {
 
     @Override
     public String getStatusText() {
-        return statusText;
+        if(status.equals("sample_data_available_accessible")) {
+            return statusText;
+        }
+        return "";
     }
 
     @Override
