+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifeCycle/requestStatus/RequestStatusStart.java
@@ -1,4 +1,4 @@
-package de.samply.bbmri.negotiator.util.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
 
 import de.samply.bbmri.negotiator.model.RequestStatusDTO;
 
@@ -6,14 +6,16 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusAbandoned implements RequestStatus {
+public class RequestStatusStart implements RequestStatus {
 
-    private String statusType = "abandoned";
-    private String statusText = "Negotiation abandoned";
+    private String status = null;
+    private String statusType = "start";
+    private String statusText = "Start Negotiation";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList();
+    private List allowedNextStatus = Arrays.asList("started", "abandoned");
 
-    public RequestStatusAbandoned(RequestStatusDTO requestStatus) {
+    public RequestStatusStart(RequestStatusDTO requestStatus) {
+        status = requestStatus.getStatus();
         statusDate = requestStatus.getStatusDate();
     }
 
@@ -24,7 +26,7 @@ public class RequestStatusAbandoned implements RequestStatus {
 
     @Override
     public String getStatus() {
-        return "abandoned";
+        return status;
     }
 
     @Override
@@ -44,7 +46,7 @@ public class RequestStatusAbandoned implements RequestStatus {
 
     @Override
     public List getAllowedNextStatus() {
-        return null;
+        return allowedNextStatus;
     }
 
     @Override
@@ -59,6 +61,6 @@ public class RequestStatusAbandoned implements RequestStatus {
 
     @Override
     public String getTableRow() {
-        return "<tr><td>" + statusDate + "</td><td>abandoned</td><td></td><td></tr>";
+        return "<tr><td>null</td><td>start</td><td></td><td></tr>";
     }
 }
