+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifeCycle/util/LifeCycleStatusUtil.java
@@ -1,0 +1,4 @@
+package eu.bbmri.eric.csit.service.negotiator.lifeCycle.util;
+
+public class LifeCycleStatusUtil {
+}
