+++ src/test/java/de/samply/bbmri/negotiator/util/requestStatus/RequestStatusReviewTest.java
@@ -1,6 +1,7 @@
 package de.samply.bbmri.negotiator.util.requestStatus;
 
 import de.samply.bbmri.negotiator.model.RequestStatusDTO;
+import eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus.RequestStatusReview;
 import org.junit.Before;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
